Re-organize XMLSettings
