+++ src/auto_splitter_settings.rs
@@ -4,86 +4,11 @@ use asr::future::retry;
 use std::{path::Path, fs::File, io::{self, Read}};
 use xmltree::{Element, XMLNode};
 
-#[derive(Clone, Debug)]
-pub struct XMLSettings {
-    name: Option<String>,
-    children: Vec<XMLNode>,
-    list_items: Vec<(String, String)>,
-}
-
-impl Default for XMLSettings {
-    fn default() -> Self { XMLSettings { name: None, children: vec![], list_items: Vec::new() } }
-}
-
-impl XMLSettings {
-    pub fn from_file<P: AsRef<Path>>(path: P, list_items: &[(&str, &str)]) -> Option<Self> {
-        let list_items = list_items.into_iter().map(|(l, i)| (l.to_string(), i.to_string())).collect();
-        let bs = file_read_all_bytes(path).ok()?;
-        let es = Element::parse_all(bs.as_slice()).ok()?;
-        let auto_splitter_settings = es.iter().find_map(xml_find_auto_splitter_settings)?;
-        Some(XMLSettings { name: None, children: auto_splitter_settings, list_items })
-    }
-    pub fn from_xml_string(s: &str, list_items: &[(&str, &str)]) -> Result<Self, xmltree::ParseError> {
-        let list_items = list_items.into_iter().map(|(l, i)| (l.to_string(), i.to_string())).collect();
-        Ok(XMLSettings { name: None, children: Element::parse_all(s.as_bytes())?, list_items })
-    }
-    fn is_list_get_item_name(&self) -> Option<&str> {
-        let n = self.name.as_deref()?;
-        for (l, i) in self.list_items.iter() {
-            if n == l {
-                return Some(i);
-            }
-        }
-        None
-    }
-
-    pub fn as_string(&self) -> Option<String> {
-        match &self.children[..] {
-            [] => Some("".to_string()),
-            [XMLNode::Text(s)] => Some(s.to_string()),
-            _ => None,
-        }
-    }
-
-    pub fn as_bool(&self) -> Option<bool> {
-        match self.as_string()?.trim() {
-            "True" => Some(true),
-            "False" => Some(false),
-            _ => None,
-        }
-    }
-
-    pub fn as_list(&self) -> Option<Vec<Self>> {
-        let i = self.is_list_get_item_name()?;
-        Some(self.children.iter().filter_map(|c| {
-            match c.as_element() {
-                Some(e) if e.name == i => {
-                    Some(XMLSettings {
-                        name: Some(e.name.clone()),
-                        children: e.children.clone(),
-                        list_items: self.list_items.clone(),
-                    })
-                },
-                _ => None,
-            }
-        }).collect())
-    }
-
-    pub fn dict_get(&self, key: &str) -> Option<Self> {
-        for c in self.children.iter() {
-            match c.as_element() {
-                Some(e) if e.name == key => {
-                    return Some(XMLSettings {
-                        name: Some(e.name.clone()),
-                        children: e.children.clone(),
-                        list_items: self.list_items.clone(),
-                    });
-                },
-                _ => (),
-            }
-        }
-        None
-    }
+pub fn file_find_auto_splitter_settings<P: AsRef<Path>>(path: P) -> Option<Vec<XMLNode>> {
+    let bs = file_read_all_bytes(path).ok()?;
+    let es = Element::parse_all(bs.as_slice()).ok()?;
+    let auto_splitter_settings = es.iter().find_map(xml_find_auto_splitter_settings)?;
+    Some(auto_splitter_settings)
 }
 
 fn xml_find_auto_splitter_settings(xml: &XMLNode) -> Option<Vec<XMLNode>> {
